move Meta to cache package
